% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw-cv.R
\name{bw_dir_cv}
\alias{bw_dir_cv}
\alias{bw_dir_lcv}
\alias{bw_dir_lscv}
\title{Cross-validation bandwidth selectors for directional data}
\usage{
bw_dir_lcv(data, h_grid = exp(seq(log(0.05), log(1.5), l = 100)), L = NULL,
  plot_it = FALSE, optim = TRUE, optim_par = 0.25, optim_lower = 0.06,
  optim_upper = 10)

bw_dir_lscv(data, h_grid = exp(seq(log(0.05), log(1.5), l = 100)),
  L = NULL, plot_it = FALSE, optim = TRUE, R_code = FALSE,
  optim_par = 0.25, optim_lower = 0.06, optim_upper = 10)
}
\arguments{
\item{data}{directional data, a matrix of size \code{c(n, q + 1)}.}

\item{h_grid}{vector of bandwidths for performing a grid search. Defaults
to \code{exp(seq(log(0.05), log(1.5), l = 100))}.}

\item{L}{kernel function. Set internally to \code{function(x) exp(-x)}
(von Mises--Fisher kernel) if \code{NULL} (default).}

\item{plot_it}{display an informative plot on the optimization's grid search?
Defaults to \code{FALSE}.}

\item{optim}{run an optimization? Defaults to \code{TRUE}. Otherwise,
a grid search on \code{h} is done. Only effective if \code{L = NULL}.}

\item{optim_par, optim_lower, optim_upper}{parameters passed to \code{par},
\code{lower}, and \code{upper} in \code{\link[stats]{optim}} when using
the \code{"L-BFGS-B"} method. Default to \code{0.25}, \code{0.06}
(to avoid numerical instabilities), and \code{10}.}

\item{R_code}{use slower R code when \code{L = NULL}? Defaults to
\code{FALSE}.}
}
\value{
A list with entries:
\itemize{
  \item{\code{h_opt}: cross-validation bandwidth.}
  \item{\code{h_grid}: \code{h_grid}, if used (otherwise \code{NULL}).}
  \item{\code{CV_opt}: minimum of the CV loss.}
  \item{\code{CV_grid}: value of the CV function at \code{h_grid}, if used
  (otherwise \code{NULL}).}
}
\code{data} is not checked to have unit norm, so the user must be careful.
When \code{L = NULL}, faster FORTRAN code is employed.
}
\description{
Likelihood and least squares cross-validation bandwidth
selectors for kernel density estimation with directional data.
}
\examples{
# Sample
n <- 25
q <- 2
set.seed(42)
samp <- rotasym::r_vMF(n = n, mu = c(1, rep(0, q)), kappa = 2)

# bw_dir_lcv
bw_dir_lcv(data = samp, optim = TRUE)$h_opt
bw_dir_lcv(data = samp, optim = FALSE, plot_it = TRUE)$h_opt
bw_dir_lcv(data = samp, L = function(x) exp(-x))$h_opt

# bw_dir_lscv
bw_dir_lscv(data = samp, optim = TRUE)$h_opt
bw_dir_lscv(data = samp, optim = FALSE, plot_it = TRUE)$h_opt
bw_dir_lscv(data = samp, optim = FALSE, R_code = TRUE)$h_opt
bw_dir_lscv(data = samp, L = function(x) exp(-x))$h_opt
}
\references{
Hall, P., Watson, G. S., and Cabrera, J. (1987). Kernel density estimation
with spherical data. \emph{Biometrika}, 74(4):751--762.
\url{https://doi.org/10.1093/biomet/74.4.751}
}
