% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirStats-package.R
\docType{package}
\name{DirStats-package}
\alias{DirStats-package}
\alias{DirStats}
\title{\code{DirStats} -- Nonparametric Methods for Directional Data}
\description{
Nonparametric kernel density estimation, bandwidth selection,
and other utilities for analyzing directional data.
}
\author{
Eduardo García-Portugués.
}
